item render progress, noise generator major fix

+++ include/item/item_stack.h
@@ -11,7 +11,7 @@ typedef struct item_stack_s {
     int16_t item_damage;
 } item_stack_t;
 
-item_stack_t item_stack_create(uint8_t item_id, uint8_t stack_size, int16_t item_damage);
+item_stack_t item_stack_create(int16_t item_id, uint8_t stack_size, int16_t item_damage);
 item_stack_t item_stack_from_nbt(nbt_base_t *nbt);
 nbt_base_t *item_stack_write_nbt(item_stack_t *item_stack, nbt_base_t *nbt);
 item_stack_t item_stack_split(item_stack_t *item_stack, uint8_t split_size);
